rewinding to before sunday

+++ srcs/cs_o_print.c
@@ -6,7 +6,7 @@
 /*   By: jnederlo <jnederlo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/16 11:03:29 by jnederlo          #+#    #+#             */
-/*   Updated: 2017/07/23 17:34:42 by jnederlo         ###   ########.fr       */
+/*   Updated: 2017/07/22 18:04:20 by jnederlo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -125,10 +125,10 @@ int		cs_lc_o_def(t_type *type, t_badge *badge, char **fmt)
 
 char	*base_less_10(unsigned long long value, int base)
 {
-	int					i;
-	unsigned long long	nb;
-	unsigned long long	nbr;
-	char				*str;
+	int				i;
+	unsigned int	nb;
+	unsigned int	nbr;
+	char			*str;
 
 	i = 0;
 	nb = value;
